+++ theme/js/components/search/menu-search-option.vue
@@ -1,0 +1,51 @@
+<template>
+  <a class="fr-nav__link unstyled" :href="link">
+    <div class="fr-grid-row fr-grid-row--middle">
+      <div aria-hidden="true" class="fr-col-auto fr-mr-2w" v-html="icon">
+      </div>
+      <i18n-t keypath="Search X in Y" class="fr-col" tag="div" v-if="query">
+        <template #query>
+          <em>{{query}}</em>
+        </template>
+        <template #type>
+          <strong>{{type}}</strong>
+        </template>
+      </i18n-t>
+      <i18n-t keypath="Start typing to search in X" class="fr-col" tag="div" v-else>
+        <template #type>
+          <strong>{{type}}</strong>
+        </template>
+      </i18n-t>
+      <div aria-hidden="true" class="fr-col-auto">
+        <span class="fr-fi-arrow-right-line fr-fi--sm" aria-hidden="true"></span>
+      </div>
+    </div>
+  </a>
+</template>
+
+<script>
+export default {
+  props: {
+    icon: {
+      type: String,
+      required: true,
+    },
+    query: {
+      type: String,
+      required: true,
+    },
+    link: {
+      type: String,
+      required: true,
+    },
+    type: {
+      type: String,
+      required: true,
+    },
+  }
+}
+</script>
+
+<style scoped>
+
+</style>