+++ theme/js/components/search/search.vue
@@ -247,6 +247,7 @@ export default {
   },
   computed: {
     // Url for doing the same search (queryString only) on the reuse page
+    // TODO : switch to composition API and useSearchUrl composable
     reuseUrl() {
       return `${config.values.reuseUrl}?q=${this.queryString}`;
     },