+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Format home page numbers [#90](https://github.com/etalab/udata-front/pull/90)
 - Let browsers decide what cursor to use [#89](https://github.com/etalab/udata-front/pull/89)
 - Replace see more button on home page with link [#91](https://github.com/etalab/udata-front/pull/91)
+- Replace Suggest with accessible combobox [#88](https://github.com/etalab/udata-front/pull/88)
 
 ## 1.2.4 (2022-03-01)